requirements.txt
Flask==0.12
flask-cors==3.0.7
gunicorn==19.6.0
Jinja2==2.9.5
numpy==1.16.2
six==1.10.0
tensorflow==1.12.0
tensorflow-gpu==1.12.0
Pillow==6.0.0
requests==2.21.0